Refactored and added * for password field
