+++ frontend/src/lib/auth/schema.ts
