+++ frontend/src/routes/login/+page.server.ts
@@ -2,7 +2,7 @@ import type { PageServerLoad, Actions } from './$types.js';
 import { fail } from '@sveltejs/kit';
 import { superValidate } from 'sveltekit-superforms';
 import { zod } from 'sveltekit-superforms/adapters';
-import { formSchema } from './schema';
+import { formSchema } from '$lib/auth/schema';
 
 export const load: PageServerLoad = async () => {
 	return {
